New Alarm

+++ sipsl-code/hdr/COMAP.h
@@ -54,8 +54,8 @@ class COMAP {
 		map<CALL_OSET*, int> call_oset_doa_state[COMAPS];
 
 		//Time to live
-		map<CALL_OSET*, unsigned long long int> call_oset_ttl[COMAPS];
-		map<CALL_OSET*, unsigned long long int> call_oset_ttl_delete[COMAPS];
+		map<CALL_OSET*, lli> call_oset_ttl[COMAPS];
+		map<CALL_OSET*, lli> call_oset_ttl_delete[COMAPS];
 
 		void setDoa(CALL_OSET*, int,int);
 		int getDoa(CALL_OSET*,int);
